[FIELDS][WL] move field typings into fields folder
