+++ src/components/fields/date-field/date-field.tsx
@@ -7,8 +7,8 @@ import { useFormContext } from "react-hook-form";
 import * as Yup from "yup";
 import { DateTimeHelper, TestHelper } from "../../../utils";
 import { useValidationConfig } from "../../../utils/hooks";
-import { IGenericFieldProps } from "../../frontend-engine/types";
 import { ERROR_MESSAGES } from "../../shared";
+import { IGenericFieldProps } from "../types";
 import { IDateFieldSchema } from "./types";
 
 const DEFAULT_DATE_FORMAT = "uuuu-MM-dd";
