+++ src/components/fields/multi-select/multi-select.tsx
@@ -3,9 +3,9 @@ import { useEffect, useState } from "react";
 import { useFormContext } from "react-hook-form";
 import useDeepCompareEffect from "use-deep-compare-effect";
 import * as Yup from "yup";
+import { IGenericFieldProps } from "..";
 import { TestHelper } from "../../../utils";
 import { useValidationConfig } from "../../../utils/hooks";
-import { IGenericFieldProps } from "../../frontend-engine";
 import { ERROR_MESSAGES } from "../../shared";
 import { ISelectOption } from "../select/types";
 import { IMultiSelectOption, IMultiSelectSchema } from "./types";
