+++ src/components/fields/range-select/types.ts
@@ -1,5 +1,6 @@
 import { InputRangeSelectProps } from "@lifesg/react-design-system/input-range-select/types";
-import { IFrontendEngineBaseFieldJsonSchema, TComponentOmitProps } from "../../frontend-engine";
+import { TComponentOmitProps } from "../../frontend-engine";
+import { IBaseFieldSchema } from "../types";
 
 export interface IRangeSelectOption {
 	label: string;
@@ -7,5 +8,5 @@ export interface IRangeSelectOption {
 }
 
 export interface IRangeSelectSchema<V = undefined>
-	extends IFrontendEngineBaseFieldJsonSchema<"range-select", V>,
+	extends IBaseFieldSchema<"range-select", V>,
 		TComponentOmitProps<InputRangeSelectProps<IRangeSelectOption, string>> {}
