+++ src/components/fields/switch/types.ts
@@ -1,12 +1,11 @@
-import { IFrontendEngineBaseFieldJsonSchema, TComponentOmitProps } from "../../frontend-engine";
+import { TComponentOmitProps } from "../../frontend-engine";
+import { IBaseFieldSchema } from "../types";
 import { ToggleProps } from "@lifesg/react-design-system/toggle";
 
 type TCustomOptions = {
 	border?: boolean | undefined;
 };
 
-export interface ISwitchSchema<V = undefined>
-	extends IFrontendEngineBaseFieldJsonSchema<"switch", V>,
-		TComponentOmitProps<ToggleProps> {
+export interface ISwitchSchema<V = undefined> extends IBaseFieldSchema<"switch", V>, TComponentOmitProps<ToggleProps> {
 	customOptions?: TCustomOptions | undefined;
 }
