+++ src/components/fields/textarea/types.ts
@@ -1,8 +1,9 @@
 import { FormTextareaProps } from "@lifesg/react-design-system/form/types";
-import { IFrontendEngineBaseFieldJsonSchema, TComponentOmitProps } from "../../frontend-engine/types";
+import { TComponentOmitProps } from "../../frontend-engine";
+import { IBaseFieldSchema } from "../types";
 
 export interface ITextareaSchema<V = undefined>
-	extends IFrontendEngineBaseFieldJsonSchema<"textarea", V>,
+	extends IBaseFieldSchema<"textarea", V>,
 		TComponentOmitProps<FormTextareaProps, "maxLength"> {
 	chipTexts?: string[] | undefined;
 	chipPosition?: "top" | "bottom" | undefined;
