+++ src/components/fields/unit-number-field/types.ts
@@ -1,9 +1,10 @@
 import { UnitNumberInputProps } from "@lifesg/react-design-system/unit-number";
-import { IFrontendEngineBaseFieldJsonSchema, TComponentOmitProps } from "../../frontend-engine";
+import { TComponentOmitProps } from "../../frontend-engine";
+import { IBaseFieldSchema } from "../types";
 
 export interface IUnitNumberFieldValidationRule {
 	unitNumberFormat?: boolean | undefined;
 }
 export interface IUnitNumberFieldSchema<V = undefined>
-	extends IFrontendEngineBaseFieldJsonSchema<"unit-number-field", V, IUnitNumberFieldValidationRule>,
+	extends IBaseFieldSchema<"unit-number-field", V, IUnitNumberFieldValidationRule>,
 		TComponentOmitProps<UnitNumberInputProps> {}
