+++ src/components/frontend-engine/types.ts
@@ -1,36 +1,9 @@
-import {
-	ControllerFieldState,
-	ControllerRenderProps,
-	UseFormReset,
-	UseFormSetValue,
-	ValidationMode,
-} from "react-hook-form";
+import { UseFormReset, UseFormSetValue, ValidationMode } from "react-hook-form";
 import { TCustomComponentSchema } from "../custom";
 import { TElementSchema } from "../elements";
 import { ISectionSchema } from "../elements/section";
-import {
-	ICheckboxGroupSchema,
-	IChipsSchema,
-	IContactFieldSchema,
-	IDateFieldSchema,
-	IEmailFieldSchema,
-	IImageUploadSchema,
-	IMultiSelectSchema,
-	INumericFieldSchema,
-	IRadioButtonGroupSchema,
-	IRangeSelectSchema,
-	IResetButtonSchema,
-	ISelectSchema,
-	ISubmitButtonSchema,
-	ISwitchSchema,
-	ITextFieldSchema,
-	ITextareaSchema,
-	ITimeFieldSchema,
-	IUnitNumberFieldSchema,
-	TDateRangeFieldSchema,
-} from "../fields";
-import { ILocationFieldSchema } from "../fields/location-field/types";
-import { IYupValidationRule, TRenderRules, TYupSchemaType } from "./yup";
+import { TFieldSchema } from "../fields";
+import { TYupSchemaType } from "./yup";
 
 // =============================================================================
 // YUP SCHEMA
@@ -87,36 +60,10 @@ export interface IFrontendEngineData<V = undefined> {
 	stripUnknown?: boolean | undefined;
 }
 
-// NOTE: add all possible schema types here except section schema
-export type TFrontendEngineFieldSchema<V = undefined> =
-	| ICheckboxGroupSchema<V>
-	| IChipsSchema<V>
-	| IContactFieldSchema<V>
-	| IDateFieldSchema<V>
-	| TDateRangeFieldSchema<V>
-	| IEmailFieldSchema<V>
-	| IImageUploadSchema<V>
-	| ILocationFieldSchema<V>
-	| IMultiSelectSchema<V>
-	| IRangeSelectSchema
-	| INumericFieldSchema<V>
-	| IRadioButtonGroupSchema<V>
-	| IResetButtonSchema
-	| ISelectSchema<V>
-	| ISubmitButtonSchema
-	| ISwitchSchema<V>
-	| ITextareaSchema<V>
-	| ITextFieldSchema<V>
-	| ITimeFieldSchema<V>
-	| IUnitNumberFieldSchema<V>
-	| TCustomComponentSchema
-	| TElementSchema;
-
 export type TFrontendEngineValues<T = any> = Record<keyof T, T[keyof T]>;
 export type TRevalidationMode = Exclude<keyof ValidationMode, "onTouched" | "all">;
 export type TValidationMode = keyof ValidationMode;
 export type TRestoreMode = "none" | "default-value" | "user-input";
-
 export type TErrorMessage = string | string[] | Record<string, string | string[]>;
 export type TErrorPayload = Record<string, TErrorMessage>;
 
@@ -169,82 +116,25 @@ export interface IFrontendEngineRef extends HTMLFormElement {
 // =============================================================================
 // JSON SCHEMA
 // =============================================================================
-// NOTE: U generic is for internal use, prevents getting overwritten by custom validation types
-export interface IFrontendEngineBaseFieldJsonSchema<T, V = undefined, U = undefined> {
-	/** defines what kind of component to be rendered */
-	uiType: T;
-	/** caption for the field */
-	label: string;
-	/** render conditions
-	 * - need to fulfil at least 1 object in array (OR condition)
-	 * - in order for an object to be valid, need to fulfil all conditions in that object (AND condition) */
-	showIf?: TRenderRules[] | undefined;
-	/** validation config, can be customised by passing generics */
-	validation?: (V | U | IYupValidationRule)[];
-	/** escape hatch for other form / frontend engines to have unsupported attributes */
-	customOptions?: Record<string, unknown> | undefined;
-}
+// contains all schema types except for sections schema
+export type TFrontendEngineFieldSchema<V = undefined> = TFieldSchema<V> | TCustomComponentSchema | TElementSchema;
 
 /**
  * JSON keys to omit from field schema when extending from other interfaces
  * - keys already defined in `IFrontendEngineBaseFieldJsonSchema` to prevent collision
  * - some inherited HTML attributes
  */
-export type TFrontendEngineFieldJsonSchemaOmitKeys =
-	| "id"
-	| "label"
-	| "validation"
-	| "uiType"
-	| "showIf"
-	| "children"
-	| "value";
+type JsonSchemaOmitKeys = "id" | "label" | "validation" | "uiType" | "showIf" | "children" | "value";
 
 // NOTE: undefined allows aggregation of keys if exists
 type UnionOptionalKeys<T = undefined> = T extends string | number | symbol
-	? TFrontendEngineFieldJsonSchemaOmitKeys | T
-	: TFrontendEngineFieldJsonSchemaOmitKeys;
+	? JsonSchemaOmitKeys | T
+	: JsonSchemaOmitKeys;
 
 /**
  * Omits clashing keys between native props and frontend engine
  */
 export type TComponentOmitProps<T, V = undefined> = Omit<T, UnionOptionalKeys<V>>;
-
-/**
- * Field types
- * - components that can contain values that can get submitted
- */
-export enum EFieldType {
-	CHECKBOX = "CheckboxGroup",
-	CHIPS = "Chips",
-	"CONTACT-FIELD" = "ContactField",
-	"DATE-FIELD" = "DateField",
-	"DATE-RANGE-FIELD" = "DateRangeField",
-	"EMAIL-FIELD" = "TextField",
-	"IMAGE-UPLOAD" = "ImageUpload",
-	"LOCATION-FIELD" = "LocationField",
-	"MULTI-SELECT" = "MultiSelect",
-	"RANGE-SELECT" = "RangeSelect",
-	"NUMERIC-FIELD" = "TextField",
-	RADIO = "RadioButtonGroup",
-	RESET = "ResetButton",
-	SELECT = "Select",
-	SUBMIT = "SubmitButton",
-	SWITCH = "Switch",
-	TEXTAREA = "Textarea",
-	"TEXT-FIELD" = "TextField",
-	"TIME-FIELD" = "TimeField",
-	"UNIT-NUMBER-FIELD" = "UnitNumberField",
-}
-
-// =============================================================================
-// FIELD PROPS
-// =============================================================================
-export interface IGenericFieldProps<T = TFrontendEngineFieldSchema>
-	extends Partial<ControllerFieldState>,
-		Partial<ControllerRenderProps> {
-	id: string;
-	schema: T;
-}
 
 // =============================================================================
 // HELPERS
